change: move brush logic into its own file
