+++ crates/brush/src/lib.rs
@@ -1,9 +1,8 @@
-extern crate core;
-
+use ::util::mouse_position::WorldMousePosition;
+use ::util::SafeUnwrap;
 use anyhow::{bail, Result};
-use assets::handle::Handle;
 use assets::storage::AssetStorage;
-use assets::{Heightmap, NormalMap, Terrain, TerrainOptions};
+use assets::{NormalMap, Terrain, TerrainOptions};
 use events::DragOnWorldView;
 use gfx::{create_linear_sampler, SharedContext};
 use glam::{Vec2, Vec3};
@@ -17,9 +16,13 @@ use phobos::{
     IncompleteCommandBuffer, PipelineStage, Sampler,
 };
 use scheduler::{EventBus, EventContext, StoredSystem, System};
-use util::mouse_position::WorldMousePosition;
-use util::SafeUnwrap;
 use world::World;
+
+use crate::brushes::height::record_update_commands;
+use crate::util::terrain_uv_at;
+
+pub mod brushes;
+pub mod util;
 
 type BrushEventReceiver = tokio::sync::mpsc::Receiver<BrushEvent>;
 type BrushEventSender = tokio::sync::mpsc::Sender<BrushEvent>;
@@ -45,117 +48,6 @@ impl System<DI> for BrushSystem {
 #[derive(Debug)]
 enum BrushEvent {
     StrokeAt(Vec3),
-}
-
-const SIZE: u32 = 256;
-
-fn record_update_normals<'q>(
-    cmd: IncompleteCommandBuffer<'q, All>,
-    uv: Vec2,
-    sampler: &Sampler,
-    heights: &Heightmap,
-    normals: &NormalMap,
-) -> Result<IncompleteCommandBuffer<'q, All>> {
-    // Transition the normal map for writing
-    let cmd = cmd.transition_image(
-        &normals.image.image.view,
-        PipelineStage::TOP_OF_PIPE,
-        PipelineStage::COMPUTE_SHADER,
-        vk::ImageLayout::SHADER_READ_ONLY_OPTIMAL,
-        vk::ImageLayout::GENERAL,
-        vk::AccessFlags2::NONE,
-        vk::AccessFlags2::MEMORY_READ | vk::AccessFlags2::MEMORY_WRITE,
-    );
-    const NORMAL_SIZE: u32 = SIZE + 8;
-    let cmd = cmd.bind_compute_pipeline("normal_recompute")?;
-    let cmd = cmd
-        .bind_storage_image(0, 0, &normals.image.image.view)?
-        .bind_sampled_image(0, 1, &heights.image.image.view, sampler)?
-        .push_constant(vk::ShaderStageFlags::COMPUTE, 0, &uv)
-        .push_constant(vk::ShaderStageFlags::COMPUTE, 8, &NORMAL_SIZE)
-        .dispatch(
-            (NORMAL_SIZE as f32 / 16.0f32).ceil() as u32,
-            (NORMAL_SIZE as f32 / 16.0f32).ceil() as u32,
-            1,
-        )?;
-    // Transition the normal map back to ShaderReadOnlyOptimal for drawing
-    let cmd = cmd.transition_image(
-        &normals.image.image.view,
-        PipelineStage::COMPUTE_SHADER,
-        PipelineStage::BOTTOM_OF_PIPE,
-        vk::ImageLayout::GENERAL,
-        vk::ImageLayout::SHADER_READ_ONLY_OPTIMAL,
-        vk::AccessFlags2::MEMORY_READ | vk::AccessFlags2::MEMORY_WRITE,
-        vk::AccessFlags2::NONE,
-    );
-    Ok(cmd)
-}
-
-fn record_height_blur<'q>(
-    cmd: IncompleteCommandBuffer<'q, All>,
-    uv: Vec2,
-    heights: &Heightmap,
-) -> Result<IncompleteCommandBuffer<'q, All>> {
-    const BLUR_SIZE: u32 = SIZE + 4;
-    let cmd = cmd.bind_compute_pipeline("blur_rect")?;
-    let cmd = cmd
-        .bind_storage_image(0, 0, &heights.image.image.view)?
-        .push_constant(vk::ShaderStageFlags::COMPUTE, 0, &uv)
-        .push_constant(vk::ShaderStageFlags::COMPUTE, 8, &BLUR_SIZE)
-        .dispatch(
-            (BLUR_SIZE as f32 / 16.0f32).ceil() as u32,
-            (BLUR_SIZE as f32 / 16.0f32).ceil() as u32,
-            1,
-        )?;
-    Ok(cmd)
-}
-
-fn record_update_commands(
-    cmd: IncompleteCommandBuffer<All>,
-    uv: Vec2,
-    sampler: &Sampler,
-    heights: &Heightmap,
-    normals: &NormalMap,
-) -> Result<CommandBuffer<All>> {
-    // We are going to write to this image in a compute shader, so submit a barrier for this first.
-    let cmd = cmd.transition_image(
-        &heights.image.image.view,
-        PipelineStage::TOP_OF_PIPE,
-        PipelineStage::COMPUTE_SHADER,
-        vk::ImageLayout::SHADER_READ_ONLY_OPTIMAL,
-        vk::ImageLayout::GENERAL,
-        vk::AccessFlags2::NONE,
-        vk::AccessFlags2::MEMORY_READ | vk::AccessFlags2::MEMORY_WRITE,
-    );
-    // Bind the pipeline we will use to update the heightmap
-    let cmd = cmd.bind_compute_pipeline("height_brush")?;
-    // Bind the image to the descriptor, push our uvs to the shader and dispatch our compute shader
-    let cmd = cmd
-        .bind_storage_image(0, 0, &heights.image.image.view)?
-        .push_constant(vk::ShaderStageFlags::COMPUTE, 0, &uv)
-        .push_constant(vk::ShaderStageFlags::COMPUTE, 8, &SIZE)
-        .dispatch(SIZE / 16, SIZE / 16, 1)?;
-    // Add a barrier to synchronize with the blur shader
-    let cmd = cmd.memory_barrier(
-        PipelineStage::COMPUTE_SHADER,
-        vk::AccessFlags2::MEMORY_READ | vk::AccessFlags2::MEMORY_WRITE,
-        PipelineStage::COMPUTE_SHADER,
-        vk::AccessFlags2::MEMORY_READ | vk::AccessFlags2::MEMORY_WRITE,
-    );
-    // let cmd = record_height_blur(cmd, uv, heights)?;
-    // Transition back to ShaderReadOnlyOptimal for drawing. This also synchronizes access to the heightmap
-    // with the normal map calculation shader
-    let cmd = cmd.transition_image(
-        &heights.image.image.view,
-        PipelineStage::COMPUTE_SHADER,
-        PipelineStage::COMPUTE_SHADER,
-        vk::ImageLayout::GENERAL,
-        vk::ImageLayout::SHADER_READ_ONLY_OPTIMAL,
-        vk::AccessFlags2::MEMORY_READ | vk::AccessFlags2::MEMORY_WRITE,
-        vk::AccessFlags2::SHADER_SAMPLED_READ,
-    );
-    let cmd = record_update_normals(cmd, uv, sampler, heights, normals)?;
-    cmd.finish()
 }
 
 fn update_heightmap(uv: Vec2, bus: &EventBus<DI>, sampler: &Sampler) -> Result<()> {
@@ -190,25 +82,6 @@ fn update_heightmap(uv: Vec2, bus: &EventBus<DI>, sampler: &Sampler) -> Result<(
     Ok(())
 }
 
-fn height_uv_at(world_pos: Vec3, options: &TerrainOptions) -> Vec2 {
-    // First compute outer bounds of the terrain mesh
-    let min_x = options.min_x();
-    let min_y = options.min_y();
-    let max_x = options.max_x();
-    let max_y = options.max_y();
-    // Then we get the length of the terrain in each dimension
-    let dx = (max_x - min_x).abs();
-    let dy = (max_y - min_y).abs();
-    // Now we can simple calculate the ratio between world_pos and the length in each dimension
-    // to get the uvs.
-    // Note that we use the z coordinate since y is up, and our terrain is in the flat plane.
-    // We will assume our terrain is properly centered. In this case, the uvs we get are
-    // in the [-0.5, 0.5] range, so we need to remap them to [0, 1]. Since this range is
-    // of the same size, we can just add 0.5
-    let uv = Vec2::new(world_pos.x / dx, world_pos.z / dy);
-    uv + 0.5
-}
-
 fn use_brush_at_position(
     bus: &EventBus<DI>,
     position: Vec3,
@@ -226,7 +99,7 @@ fn use_brush_at_position(
     // We will apply our brush mainly to the heightmap texture for now. To know how
     // to do this, we need to find the UV coordinates of the heightmap texture
     // at the position we clicked at.
-    let uv = height_uv_at(position, &world.terrain_options);
+    let uv = terrain_uv_at(position, &world.terrain_options);
     update_heightmap(uv, bus, height_sampler)?;
     Ok(())
 }
